+++ src/main/kotlin/lean4ij/language/Lean4SyntaxHighlighter.kt
@@ -1,12 +1,12 @@
 package lean4ij.language
 
+import com.google.common.io.Resources
 import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.lexer.Lexer
 import com.intellij.openapi.editor.DefaultLanguageHighlighterColors
 import com.intellij.openapi.editor.HighlighterColors.BAD_CHARACTER
-import com.intellij.openapi.editor.colors.EditorColors
 import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.openapi.editor.colors.TextAttributesKey.createTextAttributesKey
 import com.intellij.openapi.fileTypes.SyntaxHighlighter
@@ -17,7 +17,7 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiElement
 import com.intellij.psi.tree.IElementType
 import lean4ij.language.psi.TokenType
-import lean4ij.language.psi.TokenType.IDENTIFIER
+import java.nio.charset.StandardCharsets
 
 
 /**
@@ -30,6 +30,7 @@ class Lean4SyntaxHighlighter : SyntaxHighlighterBase() {
     val COMMENT: TextAttributesKey = createTextAttributesKey("LEAN_COMMENT", DefaultLanguageHighlighterColors.LINE_COMMENT)
     val NUMBER: TextAttributesKey = createTextAttributesKey("LEAN_NUMBER", DefaultLanguageHighlighterColors.NUMBER)
     val SORRY : TextAttributesKey = createTextAttributesKey("LEAN_SORRY", DefaultLanguageHighlighterColors.INVALID_STRING_ESCAPE)
+    val KEYWORD_IN_PROOF : TextAttributesKey = createTextAttributesKey("LEAN_KEYWORD_IN_PROOF", DefaultLanguageHighlighterColors.KEYWORD)
 
     val BAD_CHAR_KEYS: Array<TextAttributesKey> = arrayOf(BAD_CHARACTER)
     val SEPARATOR_KEYS: Array<TextAttributesKey> = arrayOf(SEPARATOR)
@@ -39,6 +40,7 @@ class Lean4SyntaxHighlighter : SyntaxHighlighterBase() {
     val NUMBER_KEYS: Array<TextAttributesKey> = arrayOf(NUMBER)
     val EMPTY_KEYS: Array<TextAttributesKey> = arrayOf()
     val SORRY_KEYS: Array<TextAttributesKey> = arrayOf(SORRY)
+    val KEYWORD_IN_PROOF_KEYS: Array<TextAttributesKey> = arrayOf(KEYWORD_IN_PROOF)
 
 
     override fun getHighlightingLexer(): Lexer {
@@ -59,6 +61,9 @@ class Lean4SyntaxHighlighter : SyntaxHighlighterBase() {
             tokenType == TokenType.KEYWORD_COMMAND_PREFIX
             ) {
             return KEY_KEYS;
+        }
+        if (tokenType == TokenType.KEYWORD_COMMAND6) {
+            return KEYWORD_IN_PROOF_KEYS
         }
         if (tokenType == TokenType.LINE_COMMENT || tokenType == TokenType.BLOCK_COMMENT || tokenType == TokenType.DOC_COMMENT) {
             return COMMENT_KEYS;
@@ -84,11 +89,46 @@ class Lean4SyntaxHighlighterFactory : SyntaxHighlighterFactory() {
  * TODO use customized text attributes
  */
 class Lean4Annotator : Annotator {
+
+    companion object {
+        val tactics = getAllTactics()
+
+        private fun getAllTactics(): Map<String, String> {
+            val tactics = mutableMapOf<String, String>()
+            val resource = javaClass.classLoader.getResource("tactics.txt")?:return emptyMap()
+            for (line in Resources.readLines(resource, StandardCharsets.UTF_8)) {
+                if (line.startsWith("--")) {
+                    continue
+                }
+                val (key, value) = line.split(" ")
+                tactics[key] = value
+            }
+            return tactics.toMap()
+        }
+    }
+
     override fun annotate(element: PsiElement, holder: AnnotationHolder) {
         if (element.parent is Lean4Definition) {
-            if (element.node.elementType == IDENTIFIER) {
+            if (element.node.elementType == TokenType.IDENTIFIER || element.node.elementType == TokenType.DOT) {
                 holder.newSilentAnnotation(HighlightSeverity.INFORMATION)
                     .range(element.textRange).textAttributes(DefaultLanguageHighlighterColors.FUNCTION_DECLARATION).create();
+            }
+        }
+        // check the parent rather than the element itself for skipping comments
+        if (element.parent is Lean4Attributes) {
+            if (element.node.elementType == TokenType.IDENTIFIER) {
+                holder.newSilentAnnotation(HighlightSeverity.INFORMATION)
+                    .range(element.textRange).textAttributes(DefaultLanguageHighlighterColors.METADATA).create();
+            }
+            if (element.node.elementType == TokenType.ATTRIBUTE) {
+                holder.newSilentAnnotation(HighlightSeverity.INFORMATION)
+                    .range(element.textRange).textAttributes(DefaultLanguageHighlighterColors.KEYWORD).create();
+            }
+        }
+        if (element.node.elementType == TokenType.IDENTIFIER) {
+            if (tactics.containsKey(element.text)) {
+                holder.newSilentAnnotation(HighlightSeverity.INFORMATION)
+                    .range(element.textRange).textAttributes(DefaultLanguageHighlighterColors.FUNCTION_CALL).create();
             }
         }
     }