+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - some tuning for the basic syntax highlight
 - some refactor for the setting page implementation
 - remove extra css setting in the setting page for currently messing up with theme
+- basic toolwindow toolbar for infoview
+- switch to a new icon 
 
 ## [0.1.2] - 2024-11-03